+++ luxonis_ml/data/augmentations/mosaic.py
@@ -1,53 +1,66 @@
+from typing import Any, Dict, List, Optional, Tuple, Union
+
 import numpy as np
-from typing import Tuple, Dict, Any, List, Optional
+from albumentations import BoxType, KeypointType
+from albumentations.core.bbox_utils import (
+    denormalize_bbox,
+    normalize_bbox,
+)
 from albumentations.core.transforms_interface import (
     BoxInternalType,
     ImageColorType,
     KeypointInternalType,
 )
-from albumentations.core.bbox_utils import (
-    denormalize_bbox,
-    normalize_bbox,
-)
+
 from .batch_transform import BatchBasedTransform
+from .utils import AUGMENTATIONS
 
 
+@AUGMENTATIONS.register_module()
 class Mosaic4(BatchBasedTransform):
-    """Mosaic augmentation arranges randomly selected four images into single image in a 2x2 grid layout.
-    The input images should have the same number of channels but can have different widths and heights.
-    The output is cropped around the intersection point of the four images with the size (out_with x out_height).
-    If the mosaic image is smaller than with x height, the gap is filled by the fill_value.
-    Args:
-        out_height (int)): output image height.
-            The mosaic image is cropped by this height around the mosaic center.
-            If the size of the mosaic image is smaller than this value the gap is filled by the `value`.
-        out_width (int): output image width.
-            The mosaic image is cropped by this height around the mosaic center.
-            If the size of the mosaic image is smaller than this value the gap is filled by the `value`.
-        value (int, float, list of ints, list of float): padding value. Default 0 (None).
-        replace (bool): whether to allow replacement in sampling or not. When the value is `True`, the same image
-            can be selected multiple times. When False, the batch size of the input should be at least four.
-        out_batch_size(int): output batch size. If the replace = False,
-            the input batch size should be 4 * out_batch_size.
-        mask_value (int, float, list of ints, list of float): padding value for masks. Default 0 (None).
-    Targets:
-        image_batch, mask_batch, bboxes_batch
-    [Bochkovskiy] Bochkovskiy A, Wang CY, Liao HYM. (2020) "YOLOv 4 : Optimal speed and accuracy of object detection.",
-    https://arxiv.org/pdf/2004.10934.pdf
-    """
-
     def __init__(
         self,
-        out_height,
-        out_width,
-        value=None,
-        replace=True,
-        out_batch_size=1,
-        mask_value=None,
-        always_apply=False,
-        p=0.5,
+        out_height: int,
+        out_width: int,
+        value: Optional[Union[int, float, List[int], List[float]]] = None,
+        out_batch_size: int = 1,
+        mask_value: Optional[Union[int, float, List[int], List[float]]] = None,
+        always_apply: bool = False,
+        p: float = 0.5,
     ):
-        super().__init__(always_apply=always_apply, p=p)
+        """Mosaic augmentation arranges selected four images into single image in a 2x2
+        grid layout. This is done in deterministic way meaning first image in the batch
+        will always be in top left. The input images should have the same number of
+        channels but can have different widths and heights. The output is cropped around
+        the intersection point of the four images with the size (out_with x out_height).
+        If the mosaic image is smaller than with x height, the gap is filled by the
+        fill_value.
+
+        @param out_height: Output image height. The mosaic image is cropped by this height around the mosaic center.
+        If the size of the mosaic image is smaller than this value the gap is filled by the `value`.
+        @type out_height: int
+
+        @param out_width: Output image width. The mosaic image is cropped by this height around the mosaic center.
+        If the size of the mosaic image is smaller than this value the gap is filled by the `value`.
+        @type out_width: int
+
+        @param value: Padding value. Defaults to None.
+        @type value: Optional[Union[int, float, List[int], List[float]]], optional
+
+        @param out_batch_size: Number of output images in the batch. Defaults to 1.
+        @type out_batch_size: int, optional
+
+        @param mask_value: Padding value for masks. Defaults to None.
+        @type mask_value: Optional[Union[int, float, List[int], List[float]]], optional
+
+        @param always_apply: Whether to always apply the transform. Defaults to False.
+        @type always_apply: bool, optional
+
+        @param p: Probability of applying the transform. Defaults to 0.5.
+        @type p: float, optional
+        """
+
+        super().__init__(batch_size=4, always_apply=always_apply, p=p)
 
         if out_height <= 0:
             raise ValueError(f"out_height should be larger than 0, got {out_height}")
@@ -58,15 +71,19 @@ class Mosaic4(BatchBasedTransform):
                 f"out_batch_size should be larger than 0, got {out_batch_size}"
             )
 
-        self.n_tiles = 4  # 2x2
+        self.n_tiles = self.batch_size  # 4: 2x2
         self.out_height = out_height
         self.out_width = out_width
-        self.replace = replace
         self.value = value
         self.mask_value = mask_value
         self.out_batch_size = out_batch_size
 
     def get_transform_init_args_names(self) -> Tuple[str, ...]:
+        """Gets the default arguments for the mixup augmentation.
+
+        @return: The string keywords of the arguments.
+        @rtype: Tuple[str, ...]
+        """
         return (
             "out_height",
             "out_width",
@@ -78,9 +95,28 @@ class Mosaic4(BatchBasedTransform):
 
     @property
     def targets_as_params(self):
+        """List of augmentation targets.
+
+        @return: Output list of augmentation targets.
+        @rtype: List[str]
+        """
         return ["image_batch"]
 
-    def apply_to_image_batch(self, image_batch, indices, **params):
+    def apply_to_image_batch(
+        self, image_batch: List[np.ndarray], indices: List[int], **params
+    ) -> List[np.ndarray]:
+        """Applies the transformation to a batch of images.
+
+        @param image_batch: Batch of input images to which the transformation is
+            applied.
+        @type image_batch: List[np.ndarray]
+        @param indices: Indices of images in the batch.
+        @type indices: List[Tuple[int, int]]
+        @param params: Additional parameters for the transformation.
+        @type params: Any
+        @return: List of transformed images.
+        @rtype: List[np.ndarray]
+        """
         output_batch = []
         for i_batch in range(self.out_batch_size):
             idx_chunk = indices[self.n_tiles * i_batch : self.n_tiles * (i_batch + 1)]
@@ -89,7 +125,20 @@ class Mosaic4(BatchBasedTransform):
             output_batch.append(mosaiced)
         return output_batch
 
-    def apply_to_mask_batch(self, mask_batch, indices, **params):
+    def apply_to_mask_batch(
+        self, mask_batch: List[np.ndarray], indices: List[int], **params
+    ) -> List[np.ndarray]:
+        """Applies the transformation to a batch of masks.
+
+        @param mask_batch: Batch of input masks to which the transformation is applied.
+        @type mask_batch: List[np.ndarray]
+        @param indices: Indices of images in the batch.
+        @type indices: List[Tuple[int, int]]
+        @param params: Additional parameters for the transformation.
+        @type params: Any
+        @return: List of transformed masks.
+        @rtype: List[np.ndarray]
+        """
         output_batch = []
         for i_batch in range(self.out_batch_size):
             idx_chunk = indices[self.n_tiles * i_batch : self.n_tiles * (i_batch + 1)]
@@ -100,7 +149,27 @@ class Mosaic4(BatchBasedTransform):
             output_batch.append(mosaiced)
         return output_batch
 
-    def apply_to_bboxes_batch(self, bboxes_batch, indices, image_shapes, **params):
+    def apply_to_bboxes_batch(
+        self,
+        bboxes_batch: List[BoxType],
+        indices: List[int],
+        image_shapes: List[Tuple[int, int]],
+        **params,
+    ) -> List[BoxType]:
+        """Applies the transformation to a batch of bboxes.
+
+        @param bboxes_batch: Batch of input bboxes to which the transformation is
+            applied.
+        @type bboxes_batch: List[BboxType]
+        @param indices: Indices of images in the batch.
+        @type indices: List[Tuple[int, int]]
+        @param image_shapes: Shapes of the input images in the batch.
+        @type image_shapes: List[Tuple[int, int]]
+        @param params: Additional parameters for the transformation.
+        @type params: Any
+        @return: List of transformed bboxes.
+        @rtype: List[BoxType]
+        """
         output_batch = []
         for i_batch in range(self.out_batch_size):
             idx_chunk = indices[self.n_tiles * i_batch : self.n_tiles * (i_batch + 1)]
@@ -119,8 +188,26 @@ class Mosaic4(BatchBasedTransform):
         return output_batch
 
     def apply_to_keypoints_batch(
-        self, keyboints_batch, indices, image_shapes, **params
-    ):
+        self,
+        keyboints_batch: List[KeypointType],
+        indices: List[int],
+        image_shapes: List[Tuple[int, int]],
+        **params,
+    ) -> List[KeypointType]:
+        """Applies the transformation to a batch of keypoints.
+
+        @param keypoints_batch: Batch of input keypoints to which the transformation is
+            applied.
+        @type keypoints_batch: List[KeypointType]
+        @param indices: Indices of images in the batch.
+        @type indices: List[Tuple[int, int]]
+        @param image_shapes: Shapes of the input images in the batch.
+        @type image_shapes: List[Tuple[int, int]]
+        @param params: Additional parameters for the transformation.
+        @type params: Any
+        @return: List of transformed keypoints.
+        @rtype: List[KeypointType]
+        """
         output_batch = []
         for i_batch in range(self.out_batch_size):
             idx_chunk = indices[self.n_tiles * i_batch : self.n_tiles * (i_batch + 1)]
@@ -139,15 +226,22 @@ class Mosaic4(BatchBasedTransform):
         return output_batch
 
     def get_params_dependent_on_targets(self, params: Dict[str, Any]) -> Dict[str, Any]:
+        """Get parameters dependent on the targets.
+
+        @param params: Dictionary containing parameters.
+        @type params: Dict[str, Any]
+        @return: Dictionary containing parameters dependent on the targets.
+        @rtype: Dict[str, Any]
+        """
         image_batch = params["image_batch"]
         n = len(image_batch)
-        if not self.replace and self.n_tiles * self.out_batch_size > n:
+        if self.n_tiles * self.out_batch_size > n:
             raise ValueError(
-                f"If replace == False, the batch size (= {n}) should be larger than "
+                f"The batch size (= {n}) should be larger than "
                 + f"{self.n_tiles} x out_batch_size (= {self.n_tiles * self.out_batch_size})"
             )
         indices = np.random.choice(
-            range(n), size=self.n_tiles * self.out_batch_size, replace=self.replace
+            range(n), size=self.n_tiles * self.out_batch_size, replace=False
         ).tolist()
         image_shapes = [tuple(image.shape[:2]) for image in image_batch]
         return {
@@ -168,11 +262,17 @@ def mosaic4(
     If the mosaic image is smaller than with x height, the gap is filled by the fill_value.
     This implementation is based on YOLOv5 with some modification:
     https://github.com/ultralytics/yolov5/blob/932dc78496ca532a41780335468589ad7f0147f7/utils/datasets.py#L648
-    Args:
-        image_batch (List[np.ndarray]): image list. The length should be four. Each image can has different size.
-        height (int): Height of output mosaic image
-        width (int): Width of output mosaic image
-        value (int, float, list of ints, list of float): padding value
+
+    @param image_batch: Image list. The length should be four. Each image can has different size.
+    @type image_batch: List[np.ndarray]
+    @param height: Height of output mosaic image
+    @type height: int
+    @param width: Width of output mosaic image
+    @type width: int
+    @param value: Padding value
+    @type value: Optional[ImageColorType]
+    @return: Final output image
+    @rtype: np.ndarray
     """
     N_TILES = 4
     if len(image_batch) != N_TILES:
@@ -271,13 +371,22 @@ def bbox_mosaic4(
     width: int,
 ) -> BoxInternalType:
     """Put the given bbox in one of the cells of the 2x2 grid.
-    Args:
-        bbox (tuple): A bounding box `(x_min, y_min, x_max, y_max)`.
-        rows (int): Height of input image that corresponds to one of the mosaic cells
-        cols (int): Width of input image that corresponds to one of the mosaic cells
-        position_index (int): Index of the mosaic cell. 0: top left, 1: top right, 2: bottom left, 3: bottom right
-        height (int): Height of output mosaic image
-        width (int): Width of output mosaic image
+
+    @param bbox: A bounding box `(x_min, y_min, x_max, y_max)`.
+    @type bbox: BoxInternalType
+    @param rows: Height of input image that corresponds to one of the mosaic cells
+    @type rows: int
+    @param cols: Width of input image that corresponds to one of the mosaic cells
+    @type cols: int
+    @param position_index: Index of the mosaic cell. 0: top left, 1: top right, 2:
+        bottom left, 3: bottom right
+    @type position_index: int
+    @param height: Height of output mosaic image
+    @type height: int
+    @param width: Width of output mosaic image
+    @type width: int
+    @return: Transformed bbox
+    @rtype: BoxInternalType
     """
     bbox = denormalize_bbox(bbox, rows, cols)
     center_x = width // 2
@@ -313,13 +422,22 @@ def keypoint_mosaic4(
     width: int,
 ) -> KeypointInternalType:
     """Put the given bbox in one of the cells of the 2x2 grid.
-    Args:
-        keypoint (tuple): A keypoint `(x, y, angle, scale)`.
-        rows (int): Height of input image that corresponds to one of the mosaic cells
-        cols (int): Width of input image that corresponds to one of the mosaic cells
-        position_index (int): Index of the mosaic cell. 0: top left, 1: top right, 2: bottom left, 3: bottom right
-        height (int): Height of output mosaic image
-        width (int): Width of output mosaic image
+
+    @param keypoint: A keypoint `(x, y, angle, scale)`.
+    @type bbox: KeypointInternalType
+    @param rows: Height of input image that corresponds to one of the mosaic cells
+    @type rows: int
+    @param cols: Width of input image that corresponds to one of the mosaic cells
+    @type cols: int
+    @param position_index: Index of the mosaic cell. 0: top left, 1: top right, 2:
+        bottom left, 3: bottom right
+    @type position_index: int
+    @param height: Height of output mosaic image
+    @type height: int
+    @param width: Width of output mosaic image
+    @type width: int
+    @return: Transformed keypoint
+    @rtype: KeypointInternalType
     """
     x, y, angle, scale = keypoint
